Fix batch size in inference
The dataloader in inference is using the configuration batch size. When loading a new graph line 82, the whole batch was loaded on a single graph. For a config with a batch_size !=1 this break the inference. For example, with batch_size=3, the loaded graph is : Graph(num_nodes=5769, num_edges=33210,...) with a problem of num_nodes=1923. This lead to an error line 119, where mask and pred_i[:, 0:2] have incompatible shapes.

This fix set the batch_size to one, to ensure the inference is running with only one graph loaded at a time to have compatible shapes.

Signed-off-by: Briac Malbois Le Borgne <briac.malbois_le_borgne@ens-paris-saclay.fr>